// server/types.ts
export type TaxRegime = 'SimplesNacional' | 'LucroPresumido' | 'LucroReal';

export interface Client {
  id: number;
  name: string;
  company: string;
  email: string;
  phone: string;
  status: 'Ativo' | 'Inativo';
  taxRegime: TaxRegime;
  cnaes: string[];
  keywords: string[];
  businessDescription: string;
}